+++ deeb_server/src/api/insert_many/mod.rs
@@ -42,24 +42,6 @@ pub async fn insert_many(
         }
     }
 
-    // Create Instance
-    match database
-        .deeb
-        .add_instance(
-            format!("{}-{}", &path.entity_name, app_data.instance_name.as_str()).as_str(),
-            &format!("./db/{}.json", app_data.instance_name),
-            vec![entity.clone()],
-        )
-        .await
-    {
-        Ok(_) => {}
-        Err(err) => {
-            log::error!("{:?}", err);
-            return Response::new(StatusCode::INTERNAL_SERVER_ERROR)
-                .message("Failed to get instance.");
-        }
-    };
-
     let allowed = app_data.rules_worker.check_rules(
         &crate::rules::AccessOperation::InsertMany,
         &path.entity_name,