+++ deeb_server/src/api/update_one/mod.rs
@@ -28,24 +28,6 @@ pub async fn update_one(
     let database = app_data.database.clone();
     let entity = Entity::new(&path.entity_name);
 
-    // Create Instance
-    match database
-        .deeb
-        .add_instance(
-            format!("{}-{}", &path.entity_name, app_data.instance_name.as_str()).as_str(),
-            &format!("./db/{}.json", app_data.instance_name),
-            vec![entity.clone()],
-        )
-        .await
-    {
-        Ok(_) => {}
-        Err(err) => {
-            log::error!("{:?}", err);
-            return Response::new(StatusCode::INTERNAL_SERVER_ERROR)
-                .message("Failed to get instance.");
-        }
-    };
-
     let applied_query = match app_data.rules_worker.get_query(
         &AccessOperation::UpdateOne,
         &path.entity_name,
@@ -75,7 +57,10 @@ pub async fn update_one(
         client_query
     };
 
-    let record = database.deeb.find_one::<Value>(&entity, query.clone(), None).await;
+    let record = database
+        .deeb
+        .find_one::<Value>(&entity, query.clone(), None)
+        .await;
 
     if record.is_err() {
         return Response::new(StatusCode::INTERNAL_SERVER_ERROR)